+++ build.gradle
@@ -599,7 +599,8 @@ def sonarExclusions = [
   '**uk/gov/hmcts/reform/prl/services/MiamPolicyUpgradeService.java',
   '**uk/gov/hmcts/reform/prl/services/TaskListService.java',
   '**uk/gov/hmcts/reform/prl/services/tab/alltabs/AllTabServiceImpl.java',
-  '**uk/gov/hmcts/reform/prl/controllers/PrePopulateFeeAndSolicitorNameController.java'
+  '**uk/gov/hmcts/reform/prl/controllers/PrePopulateFeeAndSolicitorNameController.java',
+  '**uk/gov/hmcts/reform/prl/services/MiamPolicyUpgradeFileUploadService.java'
 ]
 
 sonarqube {
