+++ src/main/java/uk/gov/hmcts/reform/prl/services/MiamPolicyUpgradeService.java
@@ -4,21 +4,14 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections.CollectionUtils;
-import org.apache.commons.lang3.ObjectUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
-import uk.gov.hmcts.reform.prl.models.Element;
-import uk.gov.hmcts.reform.prl.models.complextypes.DomesticAbuseEvidenceDocument;
-import uk.gov.hmcts.reform.prl.models.documents.Document;
 import uk.gov.hmcts.reform.prl.models.dto.ccd.CaseData;
 import uk.gov.hmcts.reform.prl.models.dto.ccd.MiamPolicyUpgradeDetails;
-import uk.gov.hmcts.reform.prl.services.managedocuments.ManageDocumentsService;
 
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 import static org.apache.commons.lang3.ObjectUtils.isNotEmpty;
@@ -36,7 +29,6 @@ import static uk.gov.hmcts.reform.prl.enums.miampolicyupgrade.MiamPreviousAttend
 import static uk.gov.hmcts.reform.prl.enums.miampolicyupgrade.MiamPreviousAttendanceChecklistEnum.miamPolicyUpgradePreviousAttendance_Value_2;
 import static uk.gov.hmcts.reform.prl.enums.miampolicyupgrade.TypeOfMiamAttendanceEvidenceEnum.miamAttendanceDetails;
 import static uk.gov.hmcts.reform.prl.enums.miampolicyupgrade.TypeOfMiamAttendanceEvidenceEnum.miamCertificate;
-import static uk.gov.hmcts.reform.prl.utils.ElementUtils.element;
 
 @Service
 @Slf4j
@@ -44,7 +36,6 @@ import static uk.gov.hmcts.reform.prl.utils.ElementUtils.element;
 public class MiamPolicyUpgradeService {
 
     private final ObjectMapper objectMapper;
-    private final ManageDocumentsService manageDocumentsService;
 
     public Map<String, Object> populateAmendedMiamPolicyUpgradeDetails(CallbackRequest callbackRequest) {
         log.info("initial request of populateAmendedMiamPolicyUpgradeDetails " + callbackRequest.getCaseDetails().getData());
@@ -249,60 +240,5 @@ public class MiamPolicyUpgradeService {
             isExemptionForDomesticAbuse && No.equals(caseData.getMiamPolicyUpgradeDetails().getMpuIsDomesticAbuseEvidenceProvided())
                 ? caseData.getMiamPolicyUpgradeDetails().getMpuNoDomesticAbuseEvidenceReason() : null
         );
-    }
-
-    public CaseData renameConfidentialDocumentForMiamPolicyUpgrade(CaseData caseData, String systemAuthorisation) {
-        log.info("Inside renameConfidentialDocumentForMiamPolicyUpgrade");
-        if ((caseData.getMiamPolicyUpgradeDetails().getMpuExemptionReasons().contains(domesticAbuse))
-            && Yes.equals(caseData.getMiamPolicyUpgradeDetails().getMpuIsDomesticAbuseEvidenceProvided())
-            && CollectionUtils.isNotEmpty(caseData.getMiamPolicyUpgradeDetails().getMpuDomesticAbuseEvidenceDocument())) {
-            List<Element<DomesticAbuseEvidenceDocument>> mpuConfidentialDomesticAbuseEvidenceDocument = new ArrayList<>();
-            caseData.getMiamPolicyUpgradeDetails().getMpuDomesticAbuseEvidenceDocument()
-                .stream().forEach(domesticAbuseEvidenceDocument -> {
-                    log.info("Going to append Confidential prefix for Domestic Abuse Evidence Document");
-                    Document domesticAbuseDocument = manageDocumentsService.downloadAndDeleteDocument(
-                        domesticAbuseEvidenceDocument.getValue().getDomesticAbuseDocument(), systemAuthorisation);
-                    mpuConfidentialDomesticAbuseEvidenceDocument.add(element(DomesticAbuseEvidenceDocument.builder().domesticAbuseDocument(
-                        domesticAbuseDocument).build()));
-                });
-            caseData = caseData.toBuilder()
-                .miamPolicyUpgradeDetails(caseData.getMiamPolicyUpgradeDetails()
-                                              .toBuilder()
-                                              .mpuDomesticAbuseEvidenceDocument(mpuConfidentialDomesticAbuseEvidenceDocument)
-                                              .build())
-                .build();
-        }
-        if (caseData.getMiamPolicyUpgradeDetails().getMpuExemptionReasons().contains(previousMiamAttendance)
-            && ObjectUtils.isNotEmpty(caseData.getMiamPolicyUpgradeDetails().getMpuPreviousMiamAttendanceReason())) {
-            if (miamPolicyUpgradePreviousAttendance_Value_1.equals(caseData.getMiamPolicyUpgradeDetails().getMpuPreviousMiamAttendanceReason())
-                && ObjectUtils.isNotEmpty(caseData.getMiamPolicyUpgradeDetails().getMpuDocFromDisputeResolutionProvider())) {
-                log.info("Going to append Confidential prefix previous maim attendance document for option 1");
-                Document mpuDocFromDisputeResolutionProvider = manageDocumentsService.downloadAndDeleteDocument(
-                    caseData.getMiamPolicyUpgradeDetails().getMpuDocFromDisputeResolutionProvider(),
-                    systemAuthorisation
-                );
-                caseData = caseData.toBuilder()
-                    .miamPolicyUpgradeDetails(caseData.getMiamPolicyUpgradeDetails()
-                                                  .toBuilder()
-                                                  .mpuDocFromDisputeResolutionProvider(mpuDocFromDisputeResolutionProvider)
-                                                  .build())
-                    .build();
-            } else if (miamPolicyUpgradePreviousAttendance_Value_2.equals(caseData.getMiamPolicyUpgradeDetails().getMpuPreviousMiamAttendanceReason())
-                && miamCertificate.equals(caseData.getMiamPolicyUpgradeDetails().getMpuTypeOfPreviousMiamAttendanceEvidence())
-                && ObjectUtils.isNotEmpty(caseData.getMiamPolicyUpgradeDetails().getMpuCertificateByMediator())) {
-                log.info("Going to append Confidential prefix previous maim attendance document for option 2");
-                Document mpuCertificateByMediator = manageDocumentsService.downloadAndDeleteDocument(
-                    caseData.getMiamPolicyUpgradeDetails().getMpuCertificateByMediator(),
-                    systemAuthorisation
-                );
-                caseData = caseData.toBuilder()
-                    .miamPolicyUpgradeDetails(caseData.getMiamPolicyUpgradeDetails()
-                                                  .toBuilder()
-                                                  .mpuCertificateByMediator(mpuCertificateByMediator)
-                                                  .build())
-                    .build();
-            }
-        }
-        return caseData;
     }
 }
